panfrost: Add spec@egl 1.4@egl-ext_egl_image_storage to the list of failures
The kmod abstraction now enforces that private BOs (those having an
explicit VM) can't be shared. EGL_MESA_image_dma_buf_export allows
random images to be exported, but, by the time we get the export request
we've already flagged the BO as private, and which causes such export
requests to fail.

Add the test to the failing list instead of trying to fix this case.

Suggested-by: Daniel Stone <daniel.stone@collabora.com>
Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>