panvk: Prepare for Valhall buffer views
The only difference here is the fact image attributes are gone, and
texture descriptors are used instead. We rework the code so it uses
panfrost_new_texture() to emit the texture/plane descriptors, which
leaves us with one less thing to worry about.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Acked-by: Boris Brezillon <boris.brezillon@collabora.com>
Part-of: <https://gitlab.freedesktop.org/panfrost/mesa/-/merge_requests/72>